pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axol"
version = "0.1.0"
description = "AI-native vector programming language"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]
gpu = ["cupy-cuda12x>=12.0.0"]
jax = ["jax[cpu]>=0.4.0"]
server = ["fastapi>=0.100.0", "uvicorn>=0.23.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
